requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.11.7
pydantic_core==2.33.2
